requirements.txt
accelerate==1.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.2.0
altair==5.0.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow-cpp==17.0.0
asttokens==2.0.5
async-lru==2.0.4
attrs==24.3.0
aws-c-auth==0.6.19
aws-c-cal==0.5.20
aws-c-common==0.8.5
aws-c-compression==0.2.16
aws-c-event-stream==0.2.15
aws-c-http==0.6.25
aws-c-io==0.13.10
aws-c-mqtt==0.7.13
aws-c-s3==0.1.51
aws-c-sdkutils==0.1.6
aws-checksums==0.1.13
aws-crt-cpp==0.18.16
aws-sdk-cpp==1.11.212
babel==2.11.0
backoff==2.2.1
beautifulsoup4==4.12.3
blas==1.0
bleach==6.2.0
blinker==1.6.2
boost-cpp==1.82.0
bottleneck==1.4.2
brotli-python==1.0.9
bzip2==1.0.8
c-ares==1.19.1
ca-certificates==2024.12.31
cachetools==4.2.4
certifi==2024.12.14
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.5
datasets==2.12.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
deprecated==1.2.18
dill==0.3.6
distro==1.9.0
docx2txt==0.8
ecdsa==0.19.0
effdet==0.4.1
emoji==2.14.1
eval-type-backport==0.2.2
executing==0.8.3
expat==2.6.4
faiss-cpu==1.9.0.post1
fastapi==0.112.2
filelock==3.13.1
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.56.0
freetype==2.12.1
frozenlist==1.5.0
fsspec==2024.12.0
gflags==2.2.2
gitdb==4.0.7
gitpython==3.1.43
glog==0.5.0
gmp==6.2.1
gmpy2==2.1.2
google-api-core==1.34.1
google-auth==2.38.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
google-cloud-vision==3.10.0
googleapis-common-protos==1.63.2
greenlet==3.1.1
grpc-google-iam-v1==0.14.0
grpcio==1.62.2
grpcio-status==1.48.2
h11==0.14.0
h2==4.0.0
hpack==4.0.0
hstspreload==2020.6.30
html5lib==1.1
httpcore==1.0.2
httpx==0.28.1
httpx-sse==0.4.0
huggingface_hub==0.24.6
humanfriendly==10.0
hyperframe==6.0.1
icu==73.1
idna==3.7
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.30.0
ipython_genutils==0.2.0
ipywidgets==7.8.1
jedi==0.19.2
jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jpeg==9e
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.7.2
jupyter_events==0.10.0
jupyter_server==2.14.1
jupyter_server_terminals==0.4.4
jupyterlab==4.2.5
jupyterlab_pygments==0.1.2
jupyterlab_server==2.27.3
jupyterlab_widgets==1.0.0
kiwisolver==1.4.8
krb5==1.20.1
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.30
langchain-huggingface==0.1.2
langchain-ollama==0.2.2
langchain-text-splitters==0.3.3
langdetect==1.0.9
langsmith==0.1.147
lcms2==2.16
lerc==4.0.0
libabseil==20240116.2
libboost==1.82.0
libbrotlicommon==1.0.9
libbrotlidec==1.0.9
libbrotlienc==1.0.9
libcurl==8.11.1
libcxx==14.0.6
libdeflate==1.22
libedit==3.1.20230828
libev==4.33
libevent==2.1.12
libffi==3.4.4
libgfortran==5.0.0
libgfortran5==11.3.0
libgrpc==1.62.2
libiconv==1.16
libmpdec==4.0.0
libnghttp2==1.57.0
libopenblas==0.3.21
libpng==1.6.39
libprotobuf==4.25.3
libsentencepiece==0.2.0
libsodium==1.0.18
libssh2==1.11.1
libthrift==0.15.0
libtiff==4.5.1
libwebp-base==1.3.2
libxml2==2.13.5
libxslt==1.1.41
llvm-openmp==14.0.6
lxml==5.3.0
lz4-c==1.9.4
markdown==3.7
markdown-it-py==2.2.0
markupsafe==2.1.3
marshmallow==3.19.0
matplotlib==3.10.0
matplotlib-inline==0.1.6
mdurl==0.1.0
mistune==2.0.4
mpc==1.1.0
mpfr==4.0.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.14
mypy_extensions==1.0.0
nbclient==0.8.0
nbconvert==7.16.4
nbformat==5.10.4
ncurses==6.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.2.2
notebook-shim==0.2.3
numexpr==2.10.1
numpy==1.26.4
olefile==0.47
ollama==0.4.7
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.20.1
openai==1.63.2
opencv-python==4.11.0.86
openjpeg==2.5.2
openssl==3.0.15
orc==2.0.1
orjson==3.10.14
overrides==7.4.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.4
pdf2image==1.17.0
pdfminer-six==20240706
pexpect==4.8.0
pi-heif==0.21.0
pikepdf==9.5.2
pillow==11.0.0
pip==24.2
platformdirs==3.10.0
pluggy==1.5.0
prometheus_client==0.21.0
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.2
pyarrow==17.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.8
pycparser==2.21
pydantic==2.10.3
pydantic-core==2.27.1
pydantic-settings==2.6.1
pydeck==0.8.0
pygments==2.15.1
pyopenssl==24.2.1
pyparsing==3.2.1
pypdf==4.2.0
pypdfium2==4.30.1
pysocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0post0
python-docx==1.1.2
python-dotenv==0.21.0
python-fastjsonschema==2.20.0
python-iso639==2025.2.18
python-jose==3.4.0
python-json-logger==3.2.1
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
python-tzdata==2023.3
python-xxhash==3.5.0
pytz==2024.1
pyyaml==6.0.2
pyzmq==26.2.0
rapidfuzz==3.12.1
re2==2022.04.01
readline==8.2
referencing==0.30.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
responses==0.13.3
rfc3339-validator==0.1.4
rfc3986==1.4.0
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.7.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.15.1
send2trash==1.8.2
sentence-transformers==3.3.1
sentencepiece==0.2.0
setuptools==75.1.0
six==1.16.0
smmap==4.0.0
snappy==1.2.1
sniffio==1.3.0
soupsieve==2.5
sqlalchemy==2.0.37
sqlite==3.45.3
stack_data==0.2.0
starlette==0.38.2
streamlit==1.40.1
sympy==1.13.1
tenacity==9.0.0
terminado==0.17.1
threadpoolctl==3.5.0
timm==1.0.14
tinycss2==1.2.1
tk==8.6.14
tokenizers==0.20.1
toml==0.10.2
toolz==1.0.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.18.1
tornado==6.4.2
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
typing-extensions==4.12.2
typing_inspect==0.9.0
tzdata==2024b
unstructured==0.16.23
unstructured-client==0.30.3
unstructured-inference==0.8.7
unstructured-pytesseract==0.3.13
urllib3==2.3.0
uvicorn==0.32.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.44.0
widgetsnbextension==3.6.6
wrapt==1.17.2
xlsxwriter==3.2.2
xxhash==0.8.0
xz==5.4.6
yaml==0.2.5
yarl==1.18.0
zeromq==4.3.5
zlib==1.2.13
zstd==1.5.6